Update Sub code
Instead of using indexing I just updating the existing code to mimic the Insertion function. i.e. I mark the words that were substituted by changing their "type" to SUB or RND and once all manipulations are done change it back to "word".
This solves the issue of substituting the same words more than once. But does not address the fact that first encountered match will be substituted. You mentioned that you are ok with that. Let me know